+++ backend/src/modules/Market/market.service.ts
@@ -112,6 +112,10 @@ export class MarketService {
                 continue;
             }
 
+            if (item.irisCode === "all") {
+                continue;
+            }
+
             if (item.irisCode && item.author.irisCode && item.irisCode !== item.author.irisCode) {
                 await this.marketRepo.delete(item.id);
             }
